项目调试

+++ yudao-module-erp/yudao-module-erp-biz/src/main/java/cn/iocoder/yudao/module/erp/controller/admin/product/vo/product/ErpProductPageReqVO.java
@@ -32,7 +32,7 @@ public class ErpProductPageReqVO extends PageParam {
     private String associationSupplierId;
 
     @Schema(description = "单位编号", example = "28652")
-    private Integer unitId;
+    private Long unitId;
 
     @Schema(description = "产品状态", example = "2")
     private Integer status;
