+++ yudao-module-erp/yudao-module-erp-biz/src/main/java/cn/iocoder/yudao/module/erp/dal/dataobject/product/ErpProductDO.java
@@ -47,7 +47,7 @@ public class ErpProductDO extends BaseDO {
     /**
      * 单位编号
      */
-    private Integer unitId;
+    private Long unitId;
     /**
      * 产品状态
      */
